refactor, useQueryFetcher, QueryFetcher, fix subscribe
