+++ package.json
@@ -27,7 +27,7 @@
   },
   "peerDependencies": {
     "react": ">=16.8.0",
-    "react-relay": ">=4.0.0"
+    "react-relay": ">=5.0.0"
   },
   "devDependencies": {
     "typescript": "3.5.1",
