+++ src/index.ts
@@ -18,6 +18,7 @@ export { useMutation } from "./useMutation";
 export { default as useOssFragment } from "./useOssFragment";
 export { default as usePagination } from "./usePagination";
 export { default as useRefetch } from "./useRefetch";
+export { default as useQueryFetcher } from "./useQueryFetcher";
 export { default as useRelayEnvironment } from "./useRelayEnvironment";
 export {
   default as RelayEnvironmentProvider
